pyproject.toml
[project]
name = "ai-meet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "elevenlabs>=2.18.0",
    "fastapi>=0.119.0",
    "google-genai>=1.45.0",
    "matplotlib>=3.10.7",
    "openai-whisper>=20250625",
    "pandas>=2.3.3",
    "psutil>=7.1.3",
    "python-multipart>=0.0.20",
    "python-socketio>=5.14.2",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.37.0",
]
